+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 ### Added
 
 - Link in the footer to CHANGELOG so latest changes are visible
+- Manually add audit event when printing a NOFO
 
 ### Changed